requirements.txt
pytesseract
Pillow
PyMuPDF
pandas
